SIMPLIFY-006: Consolidate Test Helpers
- Create tests/common/test_helpers.rs with unified test infrastructure
- Remove ~110 lines of duplicated setup code across test files
- Add TestEnvironment struct for comprehensive test setup
- Create handle_helpers module for file handle allocation patterns
- Create torrent_helpers module for torrent creation utilities
- Update integration_tests.rs, fuse_operations.rs, unicode_tests.rs,
  resource_tests.rs to use common helpers
- Fix pre-existing compilation errors in common modules
- Update TODO.md and CHANGELOG.md

All tests passing (200+)

+++ tests/common/fixtures.rs
@@ -123,7 +123,7 @@ pub fn deeply_nested_torrent(id: u64, depth: usize) -> TorrentInfo {
         info_hash: format!("nested{}", id),
         name: format!("Nested {} levels", depth),
         output_folder: "/downloads".to_string(),
-        file_count: Some(files.len() as u32),
+        file_count: Some(files.len()),
         files,
         piece_length: Some(262144),
     }
